+++ routes/constants.ts
@@ -1,4 +1,5 @@
 export const ROUTES = {
+  ABOUT: "/about",
   DATA_DICTIONARY: "/data-dictionaries/ncpi-dataset-catalog",
   RESEARCH_STUDIES: "/research/studies",
 };